+++ src/app/locales/en/components/file-size.js
@@ -1,0 +1,5 @@
+import header from './file-size/header';
+
+export default {
+  header,
+};